+++ .eslintrc.cjs
@@ -1,4 +1,7 @@
 module.exports = {
+    parserOptions: {
+        ecmaVersion: 2021,
+    },
     env: {
         browser: true,
         es2021: true,