+++ lint-staged.config.cjs
@@ -1,3 +1,4 @@
 module.exports = {
     'resources/assets/**/*.{scss,css,js,ts,tsx}': ['prettier --write'],
+    // 'resources/assets/**/*.{js,ts,tsx}': ['eslint'],
 };