+++ resources/views/components/dialog/drawer/drawer.blade.php
@@ -13,16 +13,16 @@
     <div {{ $attributes->class(['fixed inset-0 z-50']) }}>
         <div
             x-on:click.stop="close()"
-            class="absolute inset-0 animate-dialog-fade-in cursor-pointer bg-black/15 backdrop-blur-[2px]"
+            class="animate-dialog-fade-in absolute inset-0 cursor-pointer bg-black/15 backdrop-blur-[2px]"
         ></div>
 
-        <div class="pointer-events-none relative flex h-full w-full animate-slide-in-nav justify-end p-2">
+        <div class="animate-slide-in-nav pointer-events-none relative flex h-full w-full justify-end p-2">
             <div
                 @class([
-                    'pointer-events-auto flex w-full flex-col rounded-xl bg-white/[0.98] shadow-md ring-1 ring-grey-100 backdrop-blur-md',
-                    'xs:w-sm' => $size === 'sm',
-                    'sm:w-md' => $size === 'md',
-                    'md:w-lg' => $size === 'lg',
+                    'ring-grey-100 pointer-events-auto flex w-full flex-col rounded-xl bg-white/[0.98] shadow-md ring-1 backdrop-blur-md',
+                    'xs:w-[640px]' => $size === 'sm',
+                    'sm:w-[768px]' => $size === 'md',
+                    'md:w-[1024px]' => $size === 'lg',
                 ])
             >
                 @if ($header)