+++ resources/views/components/form/label.blade.php
@@ -5,12 +5,14 @@
 
 @if ($slot->isNotEmpty())
     <label
-        {{ $attributes->merge(['data-slot' => 'label'])->class(['form-input-label select-none' => ! isset($unset)]) }}
+        {{ $attributes->merge(['data-slot' => 'label'])->class(['inline-flex items-start gap-1 select-none' => ! isset($unset)]) }}
     >
-        {{ $slot }}
+        <span @class(['form-input-label' => ! isset($unset)])>
+            {{ $slot }}
+        </span>
 
         @if (isset($required) && $required)
-            <span class="text-secondary-400">*</span>
+            <x-chief::badge size="xs" variant="orange" class="my-0.5">Verplicht</x-chief::badge>
         @endif
     </label>
 @endif