+++ src/Assets/App/resources/components/gallery.blade.php
@@ -3,7 +3,7 @@
 @endphp
 
 <div class="space-y-6">
-    <div class="row-start-start gutter-3">
+    <div class="gutter-3 flex flex-wrap items-start justify-start">
         @forelse ($rows as $i => $asset)
             <div wire:key="gallery_{{ $i . '_' . $asset->id }}" class="w-1/2 sm:w-1/3 md:w-1/4 xl:w-1/5 2xl:w-1/6">
                 @include('chief-assets::_partials.asset-item', ['withActions' => true])