+++ src/Forms/UI/views/templates/field.blade.php
@@ -26,14 +26,17 @@
 
 <x-chief::form.fieldset wire:ignore.self :attributes="$fieldSetAttributes">
     @if ($getLabel())
-        <x-chief::form.label class="flex gap-2 items-center" :required="$isRequired()">
-            <span>{!! $getLabel() !!}</span>
-            @if($showsLocaleIndicatorInForm())
+        <div data-slot="label" class="flex items-center gap-1">
+            <x-chief::form.label :required="$isRequired()">
+                {!! $getLabel() !!}
+            </x-chief::form.label>
+
+            @if ($showsLocaleIndicatorInForm())
                 <span title="Dit veld is invulbaar per site">
-                    <x-chief::icon.locales class="size-5 text-grey-400" />
+                    <x-chief::icon.locales class="text-grey-400 size-5" />
                 </span>
             @endif
-        </x-chief::form.label>
+        </div>
     @endif
 
     @if ($getDescription())