+++ src/Fragments/UI/views/livewire/_partials/add-fragment-existing.blade.php
@@ -1,5 +1,5 @@
 <div>
-    <div class="row-start-start gutter-3">
+    <div class="gutter-3 flex flex-wrap items-start justify-start">
         <div class="w-1/2">
             <x-chief::form.fieldset>
                 <x-chief::form.label>Kies een pagina</x-chief::form.label>
@@ -24,19 +24,19 @@
         </div>
     </div>
 
-    <div data-sidebar-component="existingFragments" class="-mx-2 mt-6 space-y-2 border-t border-grey-100 pt-4">
+    <div data-sidebar-component="existingFragments" class="border-grey-100 -mx-2 mt-6 space-y-2 border-t pt-4">
         @forelse ($this->getShareableFragments() as $shareableFragment)
             <button
                 type="button"
                 wire:click="attachFragment('{{ $shareableFragment->getFragmentId() }}')"
-                class="group flex w-full items-start gap-2 rounded-xl p-2 text-left hover:bg-grey-50"
+                class="group hover:bg-grey-50 flex w-full items-start gap-2 rounded-xl p-2 text-left"
             >
-                <div class="shrink-0 text-grey-400 *:size-6 group-hover:text-primary-500">
+                <div class="text-grey-400 group-hover:text-primary-500 shrink-0 *:size-6">
                     {!! $shareableFragment->getIcon() !!}
                 </div>
 
                 <div class="grow space-y-2">
-                    <h3 class="text-base/6 font-medium text-grey-800 group-hover:text-grey-950">
+                    <h3 class="text-grey-800 group-hover:text-grey-950 text-base/6 font-medium">
                         {{ ucfirst($shareableFragment->getLabel()) }}
                     </h3>