+++ src/Fragments/UI/views/livewire/contexts.blade.php
@@ -24,7 +24,7 @@
 
     @endif
 
-    <div class="-mb-4">
+    <div>
         @foreach ($items as $item)
             <div wire:key="context-tab-content-{{ $item->id }}">
                 @if ($item->id === $activeItemId)