+++ src/Table/UI/views/components/filter/text.blade.php
@@ -5,7 +5,7 @@
 <span
     {{
         $attributes->class([
-            'btn bg-white font-normal text-grey-500 shadow-sm ring-1 ring-inset ring-grey-200 focus-within:text-grey-800 focus-within:ring-grey-300 hover:ring-grey-300',
+            'btn text-grey-500 ring-grey-200 focus-within:text-grey-800 focus-within:ring-grey-300 hover:ring-grey-300 bg-white font-normal shadow-xs ring-1 ring-inset',
             'btn-base' => $size === 'base',
             'btn-sm' => $size === 'sm',
             'btn-xs' => $size === 'xs',